pyproject.toml
[project]
name = "energyforecasting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "hypothesis>=6.142.1",
    "openmeteo-requests>=1.7.3",
    "pandas>=2.3.3",
    "pydantic>=2.12.2",
    "pydantic-settings>=2.11.0",
    "python-multipart>=0.0.20",
    "requests-cache>=1.2.1",
    "retry-requests>=2.0.0",
    "streamlit>=1.50.0",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-playwright>=0.7.1",
    "ruff>=0.14.1",
    "ty>=0.0.1a23",
]

[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py313"
respect-gitignore = true

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
docstring-code-format = true
line-ending = "auto"

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.ty.src]
exclude = ["tests"]
